+++ scripts/custom_python_prompt.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/zsh
+
+python -i -c "from sys import path; path.append('../'); from HexMeshWorkshop import *;"